+++ PowerPAN.Tests.ps1
@@ -12,8 +12,12 @@ Describe -Tag ('Unit','Acceptance') "$ModuleName Module Tests" {
          "$Here\$ModuleName.psd1" | Should -FileContentMatch "$ModuleName.psm1"
       }
 
-      It "$ModuleName class files exist in Classes directory" {
-         "$Here\Classes\*.ps1" | Should -Exist
+      It "$ModuleName class files exist in Class directory" {
+         "$Here\Class\*.ps1" | Should -Exist
+      }
+
+      It "$ModuleName Format.ps1xml files exist in Format directory" {
+         "$Here\Format\*.Format.ps1xml" | Should -Exist
       }
 
       It "$ModuleName private function files exist in Private directory" {