+++ Public/RegisteredIp/Remove-PanRegisteredIp.ps1
@@ -227,8 +227,8 @@ function Remove-PanRegisteredIp {
             Write-Debug ($MyInvocation.MyCommand.Name + ': Device: ' + $DeviceCur.Name)
             Write-Debug ($MyInvocation.MyCommand.Name + ': Prepared uid-message: ' + $CmdCur)
             $PanResponse = Invoke-PanXApi -Device $DeviceCur -Uid -Cmd $CmdCur
-            Write-Debug ($MyInvocation.MyCommand.Name + ': PanResponseStatus: ' + $PanResponse.response.status)
-            Write-Debug ($MyInvocation.MyCommand.Name + ': PanResponseMsg: ' + $PanResponse.response.InnerXml)
+            Write-Debug ($MyInvocation.MyCommand.Name + ': PanResponseStatus: ' + $PanResponse.Status)
+            Write-Debug ($MyInvocation.MyCommand.Name + ': PanResponseMsg: ' + $PanResponse.Message)
          } # else Convert $RegisteredIpAgg blueprint to XML uid-message and send API request to unregister
       } # foreach Device
    } # Process block